get rid of a few signed vs unsigned warnings

+++ cpp/OpenGL/10/TextRenderer.cpp
@@ -9,7 +9,7 @@ struct SpritePixel {
 		rgb(rgb),
 		alpha(alpha)
 	{}
-	
+
 	Vec2i pos;
 	Vec3 rgb;
 	float alpha;
@@ -19,9 +19,9 @@ void TextRenderer::render(const std::array<Vec2i,4>& tetrominoPos, const Vec3& c
 						  const std::array<Vec2i,4>& nextTetrominoPos, const Vec3& nextColor,
 						  const std::array<Vec2i,4>& targerTetrominoPos,
 						  const std::vector<Vec3>& colorDataIn) {
-							
+
 	std::vector<Vec3> colorData{colorDataIn};
-	
+
 	std::vector<SpritePixel> spritePixels;
 	for (const Vec2i& p : tetrominoPos) {
 		spritePixels.push_back(SpritePixel{p,currentColor,1});
@@ -36,7 +36,7 @@ void TextRenderer::render(const std::array<Vec2i,4>& tetrominoPos, const Vec3& c
 	}	
 
 	for (const auto& p : spritePixels) {
-		if (p.pos.y() < 0 || p.pos.x() < 0 || p.pos.y() >= height() || p.pos.x() >= width()) continue;
+		if (p.pos.y() < 0 || p.pos.x() < 0 || int64_t(p.pos.y()) >= height() || int64_t(p.pos.x()) >= width()) continue;
 		size_t index = gridIndex(p.pos.x(), p.pos.y());
 		colorData[index] = colorData.at(index) * (1-p.alpha) + p.rgb * p.alpha;
 	}
@@ -56,9 +56,9 @@ void TextRenderer::render(const std::array<Vec2i,4>& tetrominoPos, const Vec3& c
 }
 
 void TextRenderer::setColor(const Vec3& color) {
-	const uint32_t index = 16 + uint32_t(color.r()*5) + 
-							6 * uint32_t(color.g()*5) + 
+	const uint32_t index = 16 + uint32_t(color.r()*5) +
+							6 * uint32_t(color.g()*5) +
 						   36 * uint32_t(color.b()*5);
-							
+
 	std::cout << "\033[48;5;" << index << "m";
 }
